limbo: change 'append' to 'submit'
There is a patch done for a newer version of Tarantool in commit
20aad159d065bbd6e5d36bc465af8adb2bc59c30 ("limbo: handle spurious
wakeups on space waiting"). It fixes a bug not present in 3.2. But
it also changes API of the limbo.

This commit only takes the API change, but doesn't take the
bugfix. So 2.11 is as close to the master as it can be even in just
the internal APIs.

Follow-up for #11180
In scope of #11790

NO_DOC=refactoring
NO_CHANGELOG=refactoring

(cherry picked from commit 15d4744db02ed575cbb3a73200f0e3b746a8fab8)