limbo: change 'append' to 'submit'
There is a patch done for a newer version of Tarantool in commit
20aad159d065bbd6e5d36bc465af8adb2bc59c30 ("limbo: handle spurious
wakeups on space waiting"). It fixes a bug not present in 3.2. But
it also changes API of the limbo.

This commit only takes the API change, but doesn't take the
bugfix. So 3.2 is as close to the master as it can be even in just
the internal APIs.

Follow-up for #11180

NO_DOC=refactoring
NO_CHANGELOG=refactoring