codeowners: remove @Totktonada from codeowners
Alexander is going to leave the team at the end of September. Still
available for discussions, but he shouldn't be in change for the
declarative configuration.

The configuration machinery is covered by consistency tests:

* Presence of all box.cfg options in the declarative configuration.
* Presence of all compat options in the declarative configuration.
* Presence of descriptions for all the declarative configuration
  options (these descriptions are then used in the JSON Schema).

As result of this testing, I saw no significant flaws in patches sent by
other developers regarding the configuration. My most significant remark
was about using the `integer` schema node type instead of `number` where
appropriate (for example, for a value in bytes).

In all the other cases we discussed naming, feature design and so on,
not something directly related to the configuration machinery. So, I
think that I need not to assign the next person as a mandatory reviewer.

Effectively reverts commit 30fc56558132 ("Add @Totktonada as codeowner
for config").

NO_DOC=no code changes
NO_CHANGELOG=same
NO_TEST=same