limbo: fix a crash in linearization point waiting
The DB state synchronization for a linearizable transaction is
trying to wait for the receipt of all potentially confirmed
synchro txns from a remote master, and then waits for their
confirmation locally. Thus guaranteeing, that if any transaction
was committed on the master before this point, then it is now
visible on the current replica too.

Waiting for the synchro txns confirmation was done in a way that
if the limbo isn't empty, then it 100% must contain a synchro txn
in it.

But it is not always so. Sometimes it might contain a volatile
async txn, which isn't written to WAL yet. Or it might even
contain dummy entries created by the limbo flush operation (for a
snapshot, for a new replica join). About these things the
linearization sync must not care and should treat them like if the
limbo is empty.

Note that there might be more places where txn_limbo_is_empty()
seems not exactly 100% safe to use, but for none of them a
reproducer could be designed. The other places might actually be
safe until proven otherwise.

Closes #11807

NO_DOC=bugfix