journal: introduce journal_checkpoint
It was previously only available directly in WAL API. This forced
the checkpoint build code to rely on WAL, breaking the journal's
API encapsulation.

While now it is not a big deal, it is going to become one in the
next commits, where the checkpoint build code is moved into
another place not having WAL as a dependency. Lets not drag WAL
there and generalize the journal checkpoint API.

Needed for #11754

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring