box: mark replication role as system
The `replication` role (id = 3) was predefined but not included
in the system users/roles definitions. As a result, it was not
protected from being dropped or modified, unlike `guest`, `admin`,
`public`, and `super`.

This patch adds `replication` to the list of system roles and
extends `BOX_SYSTEM_USER_ID_MAX` accordingly, so that the role
is treated consistently as a system entity.

Follow-up to #1205, #3084.

Closes #11848

NO_DOC=already in doc