replication: fix wrong assumption in `box.ctl.make_bootstrap_leader`
`box_make_bootstrap_leader` assumes that by the time the instance UID is
assigned, the bootstrap strategy is also assigned. However, this assumption
does not hold during recovery, since the instance UID is assigned during
engine recovery from the snapshot at an earlier stage of recovery, and the
fiber executing `box.cfg` yields because of `iproto_do_cfg` before the
bootstrap strategy is set.

To fix this, let's throw an error iff the bootstrap strategy is set, and it
is not `supervised`.

Closes #11704

NO_DOC=<bugfix>