box: rename txn_checkpoint files
Those files were about atomic build of the "global checkpoint" of
all transaction-related control states with no observable
after-effects (nothing changed on disk and in any of those control
states).

But this code conveniently contains a lot of intricate steps which
could be reused for an on-disk checkpoint creation (snapshot).

That would turn it into something bigger than just "txn"
checkpoint. Lets prepare to that by firstly renaming the files.

The next commit will change the contents. It is separated from the
file rename in order not to confuse git into thinking that the
old txn_checkpoint files are rewritten from scratch.

Needed for #11754

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring