box: rename txn_checkpoint struct and funcs
txn_checkpoint was about atomic build of the "global checkpoint"
of all transaction-related control states with no observable
after-effects (nothing changed on disk and in any of those control
states).

But this code conveniently contains a lot of intricate steps which
could be reused for an on-disk checkpoint creation (snapshot).

That would turn it into something bigger than just "txn"
checkpoint. Lets prepare to that by renaming the struct and
related functions.

Needed for #11754

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring