box: introduce engine_checkpoint_params
It contains the data that needs to be known by the specific
engines to create their checkpoints.

Previously it was only a flag, but now it also contains the
box_checkpoint object. This one will steal some work from the
memtx checkpoint in the next commits, but its data is still going
to be written into "memtx checkpoint" aka snapshot.

Part of #11754

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring