box: collect synchro checkpoint in checkpoint.c
Previously the collection of it was done in the memtx code. But it
isn't done right and doesn't belong to just one engine.

Lets move this code into an engine-agnostic place, into
checkpoint.c. It is logically more natural and will allow to reuse
the existing code for in-memory checkpoint build.

The commit intends no functional changes and only moves the code,
so it would be easier to actually fix it in the next commit.

Part of #11754

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring

(cherry picked from commit 0a4aeb3b2ec3872a63adbbde2de37a93abaaa1ff)