box: fix outdated snapshot in presence of synchro
There was a bug that the snapshot creation code would do the
synchronous transaction control state collection too early. It
was fully done even before the engines would begin their
checkpoints.

That led to a problem that the limbo and Raft checkpoints could
have outdated confirmation LSN, terms, and anything else that is
supposed to be written into the snapshot.

The easiest way to reproduce it was to just start a snapshot when
there is at least one pending synchronous transaction. Then the
snap file would get the old confirmation LSN, but would also have
this transaction's data.

That seemed not to lead to any noticeable and visible bugs except
when one explicitly reads the snapshot file and observes the old
LSN or a term. But that still didn't look safe.

The solution is to just collect these control states the same way
as the in-memory checkpoint is being collected.

Closes #11754

NO_DOC=bugfix