box: move txn_persist_all_prepared() definition
It is going to be needed higher in the file in the next commits.
Lets move it, so the next commits would only change a small part
in it instead of having a 100% diff it due to the move.

In scope of #11754

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring

(cherry picked from commit ca7ad72765a191bca8a23c7903694e626e3f52b0)