diag: remove debug logging from functions setting diag
When we fail to write a log message to syslog to try to reconnect and
write it again in `write_to_syslog()`. If reconnecting failed and log
level is debug we log error but this involves another reconnecting
attempt and so on. As result Tarantool is crashed due to stack overflow.

Let's remove logging error in `diag_set()` and `diag_add()`. Is not
considered useful. Debug logging produces a lot of messages, to make it
helpful we are going support turning logging on by module. In this case
logging all errors in diag.h will not be helpful.

Closes #11840

NO_DOC=bugfix

(cherry picked from commit 9b3bec01ff7fc02cb5b2550d8817910ebac1d2ac)