box: introduce the 'owner' privilege
We used to have DDL operations only allowed to be performed by object
owner or an 'admin' user. This was done by explicitly checking if the
grantor ID equals `ADMIN`, so there was no way to create a custom super
user allowed to execute such a DDL. Let's introduce a new 'owner' right
and check for it instead.

Note that the 'admin' user has all possible rights set on the universe,
thus, the new one is set too: it's now owner of the universe. And such
can be any other user if the new access right is set for him.

Also, the `PRIV_GRANT` and `PRIV_REVOKE` had been introduced for pretty
logging of an error on a failed attempt to grant or revoke a privilege,
and they're not exported in Lua directly, so it must be safe to update
their values. Also, there was no documented way to allow granting a
role that had been created by another user. This problem is fixed now
as the 'owner' privilege allows to do that.

Updated all appropriate access checks for consistency by the way.

Closes #11528

@TarantoolBot document
Title: Introduce the 'owner' privilege
Since: 3.5.1

The new privilege allows a non-admin user manipulate DDL objects
created by other users the ways allowed to object creators.

Root document: https://www.tarantool.io/en/doc/latest/admin/access_control/#privileges

* The `owner` permission granted to an object or object class allows
  performing all the operations allowed to be performed by the object
  owner with the specified object or object class. Owner of an object
  class owns all objects of that class (e. g. 'space' owners own all
  spaces). Owner of the  'universe' owns all objects of all classes.