box: allow to grant role and privileges simultaneously
We used to have two exclusive paths for building the effective access
rights: after either granting a role or a privilege. That was OK since
the only useful operation could be granted to a non-creator of a role
is 'execute'. But we're going to introduce more access rights in the
future, and that will require the ability to set multiple access rights
on a role at once. This patch implements the ability.

This also fixes a case where access rights could remain granted after
revoke of last privileges from a role since having no access had been
interpreted as revoking the role, which led to not resetting grantee's
access rights on the role. That would be useless as mentioned above,
but that's a bug nevertheless.

NO_DOC=internal
NO_CHANGELOG=internal

Needed for #11528
Closes tarantool/security#150