box: read snapshot in a distinct thread to improve recovery performance
Validating requests MsgPack and parsing the requests body turns out to
be significant part of recovery in terms of CPU. Let's move it to
a distinct thread. This is similar to iproto and applier where we have
a distinct thread(s) for that.

Snapshot is written in batches of about 128k
(`XLOG_TX_AUTOCOMMIT_THRESHOLD`). The recovery thread reads the batch,
parses its requests and send to TX to process. The maximum number of
read but not processed batches is limited to 2 to make a flow control.

Closes #11888
Part of #11452

============================
= Performance test results =
============================

Recovery performance is evaluated using below perf test. The 1.10.5
performance is taken from the previous evaluations for the issue.

```
tarantool perf/lua/recovery.lua \
  --row_count 10000000 --column_count 10 --recovery_count 5 \
  --preload_script "require('compat').box_recovery_triggers_deprecation = 'new'"
```

Run                 | Perf, krps |
---------------------------------|
master, current     |    1060    |
---------------------------------|
master, patched     |    1800    |
---------------------------------|
1.10.5              |    1315    |

So introducing snapshot reader thread gives 70% boost for recovery from
snapshot. Now master is 37% faster than 1.10.5.

If `box_recovery_triggers_deprecation = 'old'` then performance is only
990 krps.

NO_TEST=performance
NO_DOC=performance