box: make the `PRIV_GRANT` work as it should
The `PRIV_GRANT` (along with `PRIV_REVOKE`) had been introduced for a
single reason: for pretty-printing the access denial error message.
Let's make it a working privilege: a user can grant accesses to an
object if it has the `PRIV_GRANT` access to it or to corresponding
entity or the universe. Dropped special admin checks from `alter.cc`
by the way as now the admin has the required ("grant") access set.

Moved grantor checks in the `priv_def_check` function into a separate
`access_check_ddl_cr` call, so now the same routine is used to check
both grantor and the current user accesses to perform grant/revoke.
Now the grantor is checked not only on object access grant/revoke,
but also for entity and universe accesses, this makes it possible to
not only grant these by admin, but also by any user with the "grant"
privilege on the entity/universe.

There's a potential problem with the new access: allowing to grant
privileges not only to admins and owners, but also the ones having
the "grant" privilege creates the following possibility:
1. We have a user "u" that has "grant" privilege on the space "s", so
   he can grant accesses to it.
2. He has granted oneself the space privilege (e. g. write to space
   "s").
3. We drop the user. Here we revoke all privileges of the user, and
   if we happen to drop the "grant space 's'" privilege first and then
   proceed to drop the "write space 's'", the latter will fail as the
   user has no "grant" privilege on the space anymore, but he is the
   grantor of the privilege and is checked.

The issue occurs on any case we check the grantor in. Let's fix it by
disallowing to grant privileges to oneself in case the privilege is
only allowed to be granted by the "grant" access, so there's no way
to drop a privilege that would solely allow to drop another one. This
also makes the "grant" privilege granular: one can grant the object to
others if it granted, but can't grant oneself more privileges using it.

Since granting a privilege to an objects requires the ability to see
it, introduced corresponding checks in all sysview spaces so that if
an object is grantable, it's visible. Except for the `_priv` space:
even if one has the "grant" privilege on the "space" entity, it can't
see space grants of other users, only the ones created by oneself.

Dropped the `PRIV_REVOKE` constant by the way, as `PRIV_GRANT` does
cover the privilege revoke too.

Closes #11528

@TarantoolBot document
Title: Introduce the "grant" privilege
Since: 3.5.1

The new privilege allows a user to grant and revoke accesses to the
object or object class the access is granted for.

Root document: https://www.tarantool.io/en/doc/latest/admin/access_control/#privileges

* The `grant` permission granted to an object or object class allows
  a non-owner of the object or non-admin user to grant and revoke
  permissions to this object or object class to other available users.
  No granting oneself is allowed though.