box: make the `PRIV_GRANT` work as it should
The `PRIV_GRANT` (along with `PRIV_REVOKE`) had been introduced for a
single reason: for pretty-printing the access denial error message.
Let's make it a working privilege: a user can grant accesses to an
object if it has the `PRIV_GRANT` access to it or to corresponding
entity or the universe.

Since granting a privilege to an objects requires the ability to see
it, introduced corresponding checks in all sysview spaces so that if
an object is grantable, it's visible. Except for the `_priv` space:
even if one has the 'grant` privilege on the 'space' entity, it can't
see space grants of other users, only the ones created by oneself.

Dropped the admin check for an entity access from the `priv_def_check`
as now it can be done in the called `access_check_ddl` along with the
new grant check.

Dropped the `PRIV_REVOKE` constant by the way, as `PRIV_GRANT` does
cover the privilege revoke too.

Closes #11528