box: allow to grant role and privileges simultaneously
We used to have two exclusive paths for building the effective access
rights: after either granting a role or a privilege. So in case of a
drop of last privileges from a role access rights could remain granted
since having no access had been interpreted as revoking the role, which
led to not resetting grantee's access rights on the role. Let's fix it
by allowing to grant a privilege and a role in one run.

Let's also panic on an error while rebuilding user grants as these can
only occur on an invalid schema or an iterator creation failure, which
are improbable in practice. So there's one less point of failure while
applying privilege grants and thus, no need to care about rolling back
the `priv_grant` or `role_grant` changes.

Also, moved the "revoke universe from admin" check to `priv_def_check`
that it seems to belong to. This required the function to be aware if
the check is performed during the `_priv` entry drop, thus, the new
`on_revoke` parameter.

Closes tarantool/security#150
Closes tarantool/security#152

NO_DOC=internal
NO_CHANGELOG=internal