uri: add the `interface` field
Needed for #11803

@TarantoolBot document
Title: `interface` field in URI table

Now `uri.parse()` can parse a table containing the `interface` field.
The value from the field is meant to be used as a name of the network
interface to bind requests to the uri to. Example: `interface = 'lo'`
to use the loopback network device or `interface = 'enp0s31f6'` to
use the local NIC specified.

The `uri.parse_many()` supports a new `default_interface` parameter
that can be used to specify the default interface to be used for uris
in the set given.