config: introduce an `iproto.ssl` section
This patch introduces a new `iproto.ssl` configuration section that can
be used to configure SSL. Previously, SSL might be configured by
providing `params` section in URIs as mentioned in the documentation
[^1]. The problem is that these URIs have been straightforwardly taken
from the configuration and instances tried to use other instances SSL
private keys to connect to them for replication and when using some of
the builtin modules such as `experimental.connpool`. The details are
available within #12030.

This fix aims to fix this problem. It introduces a new section `ssl`
that basically contains all the information the instance will use to
auth to other instances when using SSL. It uses the schema that
resembles existing `config.etcd.ssl`. Using this SSL works as intended.
Example of the new config.

```yaml
storage-1:
  iproto:
    listen:
      - uri: localhost:3301
        params:
          transport: 'ssl'
    ssl:
      ca_file: './certs/rootCA.pem'
      ssl_cert: './certs/storage-1-crt.pem'
      ssl_key: './certs/storage-1-key.pem'
storage-2:
  iproto:
    listen:
      - uri: localhost:3302
        params:
          transport: 'ssl'
    ssl:
      ca_file: './certs/rootCA.pem'
      ssl_cert: './certs/storage-2-crt.pem'
      ssl_key: './certs/storage-2-key.pem'
```

This configuration makes `storage-1` use the `storage-1` key/cert to set
the IProto socket up and to connect to `storage-2`. And `storage-2` uses
the `storage-2` key/cert to set its socket and to connect to `storage-1`
as it is expected.

Note that the old `param` section is still available. It will overwrite
less explicit behavior of the `ssl` section in favor of the backward
compatibility. Also, the parameters are not supplied to the connections
not using `ssl` as transport.

Unfortunately, using old `param` section is not alerted yet since using
configuration alert system will be introduced later since it requires
some non-trivial decisions like possible integrations with compat module
or suppressing warnings. The related issue about deprecated
configuration option is #12033.

Closes #12030
Closes tarantool/tarantool-ee#1504

NOWRAP
[^1] https://www.tarantool.io/en/doc/latest/reference/configuration/configuration_reference/#uri-params
NOWRAP

@TarantoolBot document
Title: config: new `iproto.ssl` section

A new `iproto.ssl` section has been introduced. It has the following
options.

* `iproto.ssl`

SSL parameters required for encrypted connections. These parameters
would be used to set up SSL IProto sockets and to connect to other
instances which require certificate authority (CA).

* `iproto.ssl.ca_file`

(Optional) A path to a trusted certificate authorities (CA) file. If not
set, the peer won't be checked for authenticity.

Both a server and a client can use the ca_file parameter:

- If it's on the server side, the server verifies the client.
- If it's on the client side, the client verifies the server.
- If both sides have the CA files, the server and the client verify each
  other.

* `iproto.ssl.ssl_cert`

A path to an SSL certificate file:

- For a server, it's mandatory.
- For a client, it's mandatory if the ca_file parameter is set for a
  server; otherwise, optional.

* `iproto.ssl.ssl_ciphers` (Optional) A colon-separated (:) list of SSL
  cipher suites the connection can use. Note that the list is not
  validated: if a cipher suite is unknown, Tarantool ignores it, doesn't
  establish the connection, and writes to the log that no shared cipher
  was found.

* `iproto.ssl.ssl_key` A path to a private SSL key file:

- For a server, it's mandatory.
- For a client, it's mandatory if the `ca_file` parameter is set for a
  server; otherwise, optional.

If the private key is encrypted, provide a password for it in the
`ssl_password` or `ssl_password_file` parameter

* `iproto.ssl.ssl_password`

(Optional) A password for an encrypted private SSL key provided using
`ssl_key`. Alternatively, the password can be provided in
`ssl_password_file`.

Tarantool applies the `ssl_password` and `ssl_password_file` parameters
in the following order:

- If `ssl_password` is provided, Tarantool tries to decrypt the private
  key with it.
- If `ssl_password` is incorrect or isn't provided, Tarantool tries all
  passwords from `ssl_password_file` one by one in the order they are
  written.
- If `ssl_password` and all passwords from `ssl_password_file` are
  incorrect, or none of them is provided, Tarantool treats the private
  key as unencrypted.

* `iproto.ssl.ssl_password_file` (Optional) A text file with one or more
  passwords for encrypted private SSL keys provided using `ssl_key`
  (each on a separate line). Alternatively, the password can be provided
  in `ssl_password`.

This section is a replacement for the parameters
`iproto.listen.*.params.ssl_*` and `iproto.advertise.*.params.ssl_*`.

The hints that redirect `iproto.listen.*.params.ssl_*` and
`iproto.advertise.*.params.ssl_*` to `<uri>.params.ssl_*` should be
marked as ones overwriting `iproto.ssl` options and marked for advanced
use only with a hint that the user is likely want to use `iproto.ssl`.
Also, SSL Tarantool examples should be changed w.r.t. new section
`iproto.ssl` instead.

(cherry picked from commit a54a406dd26a2ccd3072c37dca0547214c819894)