box: support `interface` URI option in `box.cfg.replication`
Part of #11803

@TarantoolBot document
Title: introduced the `interface` URI option support in `replication`

Now a URI passed to the `box.cfg.replication` can contain `interface`
option specifying the name of the interface to bind the connections
to the URI to. The `default_interface` option can be used to specify
the default `interface` value if multiple URIs specified.

Examples of the `box.cfg.replication` value:
- `{3301, interface = 'enp0s31f6'}`
- `{3301, 3302, default_interface = 'enp0s31f6'}`
- `{3301, {3302, interface = 'lo'}, default_interface = 'enp0s31f6'}