memtx: set result tuple for multikey index delete
Currently, we do not set the result tuple for multikey index deletes (i.e.,
`new_tuple == NULL`). The result tuple from multikey index replaces has
never been used, since MVCC is not supported for it, so this did not make
any difference.

In scope of #11732, we want to unify the memtx replace code paths with and
without MVCC enabled to simplify the code. To achieve this, we will save
the result tuples of index replaces to pass them down for MVCC logic and to
rollback the physical index state in case of failure.

To make the multikey index delete operation consistent with other index
operations, let's set the result tuple for deletes in the same way we do
for ordinary indexes.

Needed for #11732

NO_CHANGELOG=<refactoring>
NO_TEST=<refactoring>
NO_DOC=<refactoring>