box: don't use admin privileges in space creation
It is required in the documentation for a user to have read permission
in the `_space` to create new spaces. However, we didn't check for this.
In the patch, the behavior was changed so new id generation is executed
with user's credentials. This change allows us to further unify the id
generation logic between `_space` and `_func`.

Part of #11849
Part of #11851

NO_DOC=minor
NO_CHANGELOG=minor